requirements.txt
flask==3.0.3
gunicorn==23.0.0

numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-surprise==1.1.3
joblib>=1.3.0

google-cloud-storage>=2.0.0
